Added matmul, fixed config for qsort, matmul
