+++ .gitignore
@@ -3,3 +3,11 @@
 
 
 .idea/
+
+
+bencher/venv/
+
+bencher/tests/c/qsort/dataset.txt
+
+
+*.elf
