+++ bencher/clean_all_docker.sh
