+++ bencher/tests/c/qsort/gen_data.h
@@ -1,0 +1,7 @@
+#include <stdio.h>
+#include <stdlib.h>
+#include <time.h>
+
+
+
+void write_data(char* file_name, int* data, size_t size);
