--- bencher/tests/qsort/c/gen_data.h
@@ -1,7 +1,0 @@
-#include <stdio.h>
-#include <stdlib.h>
-#include <time.h>
-
-
-
-void write_data(char* file_name, int* data, size_t size);
